Start trying to be smarter with multiple constructors
GHC creates the :+: branches on left and right, so try and squash them
down efficiently to use just one Word8 value.

Unfortunately, this only partially works; it takes us from:

```
λ> mapM_ (print . B.unpack . ravel) [minBound .. maxBound :: LotsOfConstructors]
[0,0,0]
[0,0,1]
[0,1]
[0,2,0]
[0,2,1]
[1,0]
[1,1]
[2,0]
[2,1]
[2,2]
```

to:

```
λ> mapM_ (print . B.unpack . ravel) [minBound .. maxBound :: LotsOfConstructors]
[0,0]
[0,1]
[1,0]
[1,1]
[1,2]
[2,0]
[2,1]
[3,0]
[3,1]
[3,2]
```

This is better, but still needs work.  It looks like the instances
aren't nesting properly.

Also remove the Metadata (:+:) instance as it doesn't seem to be used;
if it was used previously then it was because of a different version
of GHC (so this should be tested).